backend/requirements.txt
Flask
requests
python-dotenv
google-genai
flask-cors
flask-socketio
websockets
